Remove unused model picker code
