+++ src/vs/workbench/contrib/chat/browser/widget/input/modelPickerActionItem.ts
@@ -3,27 +3,21 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as dom from '../../../../../../base/browser/dom.js';
-import { IActionProvider } from '../../../../../../base/browser/ui/dropdown/dropdown.js';
+import { getActiveWindow } from '../../../../../../base/browser/dom.js';
 import { IManagedHoverContent } from '../../../../../../base/browser/ui/hover/hover.js';
-import { renderIcon, renderLabelWithIcons } from '../../../../../../base/browser/ui/iconLabel/iconLabels.js';
+import { getBaseLayerHoverDelegate } from '../../../../../../base/browser/ui/hover/hoverDelegate2.js';
+import { getDefaultHoverDelegate } from '../../../../../../base/browser/ui/hover/hoverDelegateFactory.js';
+import { BaseActionViewItem } from '../../../../../../base/browser/ui/actionbar/actionViewItems.js';
 import { IAction } from '../../../../../../base/common/actions.js';
-import { IDisposable } from '../../../../../../base/common/lifecycle.js';
+import { MutableDisposable } from '../../../../../../base/common/lifecycle.js';
 import { autorun, IObservable } from '../../../../../../base/common/observable.js';
 import { localize } from '../../../../../../nls.js';
-import { IActionWidgetService } from '../../../../../../platform/actionWidget/browser/actionWidget.js';
-import { IActionWidgetDropdownAction, IActionWidgetDropdownActionProvider, IActionWidgetDropdownOptions } from '../../../../../../platform/actionWidget/browser/actionWidgetDropdown.js';
-import { ICommandService } from '../../../../../../platform/commands/common/commands.js';
 import { IContextKeyService } from '../../../../../../platform/contextkey/common/contextkey.js';
+import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
 import { IKeybindingService } from '../../../../../../platform/keybinding/common/keybinding.js';
-import { IProductService } from '../../../../../../platform/product/common/productService.js';
-import { ITelemetryService } from '../../../../../../platform/telemetry/common/telemetry.js';
-import { TelemetryTrustedValue } from '../../../../../../platform/telemetry/common/telemetryUtils.js';
-import { ChatEntitlement, IChatEntitlementService } from '../../../../../services/chat/common/chatEntitlementService.js';
-import { MANAGE_CHAT_COMMAND_ID } from '../../../common/constants.js';
 import { ILanguageModelChatMetadataAndIdentifier } from '../../../common/languageModels.js';
-import { DEFAULT_MODEL_PICKER_CATEGORY } from '../../../common/widget/input/modelPickerWidget.js';
-import { ChatInputPickerActionViewItem, IChatInputPickerOptions } from './chatInputPickerActionItem.js';
+import { IChatInputPickerOptions } from './chatInputPickerActionItem.js';
+import { ModelPickerWidget } from './chatModelPicker.js';
 
 export interface IModelPickerDelegate {
 	readonly currentModel: IObservable<ILanguageModelChatMetadataAndIdentifier | undefined>;
@@ -35,189 +29,94 @@ export interface IModelPickerDelegate {
 	showFeatured(): boolean;
 }
 
-type ChatModelChangeClassification = {
-	owner: 'lramos15';
-	comment: 'Reporting when the model picker is switched';
-	fromModel?: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The previous chat model' };
-	toModel: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The new chat model' };
-};
-
-type ChatModelChangeEvent = {
-	fromModel: string | TelemetryTrustedValue<string> | undefined;
-	toModel: string | TelemetryTrustedValue<string>;
-};
-
-
-function modelDelegateToWidgetActionsProvider(delegate: IModelPickerDelegate, telemetryService: ITelemetryService, pickerOptions: IChatInputPickerOptions): IActionWidgetDropdownActionProvider {
-	return {
-		getActions: () => {
-			const models = delegate.getModels();
-			if (models.length === 0) {
-				// Show a fake "Auto" entry when no models are available
-				return [{
-					id: 'auto',
-					enabled: true,
-					checked: true,
-					category: DEFAULT_MODEL_PICKER_CATEGORY,
-					class: undefined,
-					tooltip: localize('chat.modelPicker.auto', "Auto"),
-					label: localize('chat.modelPicker.auto', "Auto"),
-					hover: { content: localize('chat.modelPicker.auto.description', "Automatically selects the best model for your task based on capacity."), position: pickerOptions.hoverPosition },
-					run: () => { }
-				} satisfies IActionWidgetDropdownAction];
-			}
-			return models.map(model => {
-				const hoverContent = model.metadata.tooltip;
-				return {
-					id: model.metadata.id,
-					enabled: true,
-					icon: model.metadata.statusIcon,
-					checked: model.identifier === delegate.currentModel.get()?.identifier,
-					category: model.metadata.modelPickerCategory || DEFAULT_MODEL_PICKER_CATEGORY,
-					class: undefined,
-					description: model.metadata.multiplier ?? model.metadata.detail,
-					tooltip: hoverContent ? '' : model.metadata.name,
-					hover: hoverContent ? { content: hoverContent, position: pickerOptions.hoverPosition } : undefined,
-					label: model.metadata.name,
-					run: () => {
-						const previousModel = delegate.currentModel.get();
-						telemetryService.publicLog2<ChatModelChangeEvent, ChatModelChangeClassification>('chat.modelChange', {
-							fromModel: previousModel?.metadata.vendor === 'copilot' ? new TelemetryTrustedValue(previousModel.identifier) : 'unknown',
-							toModel: model.metadata.vendor === 'copilot' ? new TelemetryTrustedValue(model.identifier) : 'unknown'
-						});
-						delegate.setModel(model);
-					}
-				} satisfies IActionWidgetDropdownAction;
-			});
-		}
-	};
-}
-
-function getModelPickerActionBarActionProvider(commandService: ICommandService, chatEntitlementService: IChatEntitlementService, productService: IProductService): IActionProvider {
-
-	const actionProvider: IActionProvider = {
-		getActions: () => {
-			const additionalActions: IAction[] = [];
-			if (
-				chatEntitlementService.entitlement === ChatEntitlement.Free ||
-				chatEntitlementService.entitlement === ChatEntitlement.EDU ||
-				chatEntitlementService.entitlement === ChatEntitlement.Pro ||
-				chatEntitlementService.entitlement === ChatEntitlement.ProPlus ||
-				chatEntitlementService.entitlement === ChatEntitlement.Business ||
-				chatEntitlementService.entitlement === ChatEntitlement.Enterprise ||
-				chatEntitlementService.isInternal
-			) {
-				additionalActions.push({
-					id: 'manageModels',
-					label: localize('chat.manageModels', "Manage Models..."),
-					enabled: true,
-					tooltip: localize('chat.manageModels.tooltip', "Manage Language Models"),
-					class: undefined,
-					run: () => {
-						commandService.executeCommand(MANAGE_CHAT_COMMAND_ID);
-					}
-				});
-			}
-
-			// Add sign-in / upgrade option if entitlement is anonymous / free / new user
-			const isNewOrAnonymousUser = !chatEntitlementService.sentiment.completed ||
-				chatEntitlementService.entitlement === ChatEntitlement.Available ||
-				chatEntitlementService.anonymous ||
-				chatEntitlementService.entitlement === ChatEntitlement.Unknown;
-			if (isNewOrAnonymousUser || chatEntitlementService.entitlement === ChatEntitlement.Free) {
-				additionalActions.push({
-					id: 'moreModels',
-					label: isNewOrAnonymousUser ? localize('chat.moreModels', "Add Language Models") : localize('chat.morePremiumModels', "Add Premium Models"),
-					enabled: true,
-					tooltip: isNewOrAnonymousUser ? localize('chat.moreModels.tooltip', "Add Language Models") : localize('chat.morePremiumModels.tooltip', "Add Premium Models"),
-					class: undefined,
-					run: () => {
-						const commandId = isNewOrAnonymousUser ? 'workbench.action.chat.triggerSetup' : 'workbench.action.chat.upgradePlan';
-						commandService.executeCommand(commandId);
-					}
-				});
-			}
-
-			return additionalActions;
-		}
-	};
-	return actionProvider;
-}
-
 /**
  * Action view item for selecting a language model in the chat interface.
+ *
+ * Wraps a {@link ModelPickerWidget} and adapts it for use in an action bar,
+ * providing curated model suggestions, upgrade prompts, and grouped layout.
  */
-export class ModelPickerActionItem extends ChatInputPickerActionViewItem {
-	protected currentModel: ILanguageModelChatMetadataAndIdentifier | undefined;
+export class ModelPickerActionItem extends BaseActionViewItem {
+	private readonly _pickerWidget: ModelPickerWidget;
+	private readonly _managedHover = this._register(new MutableDisposable());
 
 	constructor(
 		action: IAction,
-		widgetOptions: Omit<IActionWidgetDropdownOptions, 'label' | 'labelRenderer'> | undefined,
 		delegate: IModelPickerDelegate,
-		pickerOptions: IChatInputPickerOptions,
-		@IActionWidgetService actionWidgetService: IActionWidgetService,
-		@IContextKeyService contextKeyService: IContextKeyService,
-		@ICommandService commandService: ICommandService,
-		@IChatEntitlementService chatEntitlementService: IChatEntitlementService,
-		@IKeybindingService keybindingService: IKeybindingService,
-		@ITelemetryService telemetryService: ITelemetryService,
-		@IProductService productService: IProductService,
+		private readonly pickerOptions: IChatInputPickerOptions,
+		@IInstantiationService instantiationService: IInstantiationService,
+		@IContextKeyService private readonly _contextKeyService: IContextKeyService,
+		@IKeybindingService private readonly keybindingService: IKeybindingService,
 	) {
-		// Modify the original action with a different label and make it show the current model
-		const actionWithLabel: IAction = {
-			...action,
-			label: delegate.currentModel.get()?.metadata.name ?? localize('chat.modelPicker.auto', "Auto"),
-			run: () => { }
-		};
+		super(undefined, action);
 
-		const baseActionBarActionProvider = getModelPickerActionBarActionProvider(commandService, chatEntitlementService, productService);
-		const modelPickerActionWidgetOptions: Omit<IActionWidgetDropdownOptions, 'label' | 'labelRenderer'> = {
-			actionProvider: modelDelegateToWidgetActionsProvider(delegate, telemetryService, pickerOptions),
-			actionBarActionProvider: { getActions: () => baseActionBarActionProvider.getActions() },
-			reporter: { id: 'ChatModelPicker', name: 'ChatModelPicker', includeOptions: true },
-		};
+		this._pickerWidget = this._register(instantiationService.createInstance(ModelPickerWidget, delegate, pickerOptions.hoverPosition));
+		this._pickerWidget.setSelectedModel(delegate.currentModel.get());
+		this._pickerWidget.setHideChevrons(pickerOptions.hideChevrons);
 
-		super(actionWithLabel, widgetOptions ?? modelPickerActionWidgetOptions, pickerOptions, actionWidgetService, keybindingService, contextKeyService, telemetryService);
-		this.currentModel = delegate.currentModel.get();
-
-		// Listen for model changes from the delegate
+		// Sync delegate → widget when model list or selection changes externally
 		this._register(autorun(t => {
 			const model = delegate.currentModel.read(t);
-			this.currentModel = model;
-			this.updateTooltip();
-			if (this.element) {
-				this.renderLabel(this.element);
-			}
+			this._pickerWidget.setSelectedModel(model);
+			this._updateTooltip();
 		}));
+
+		// Sync widget → delegate when user picks a model
+		this._register(this._pickerWidget.onDidChangeSelection(model => delegate.setModel(model)));
 	}
 
-	protected override getHoverContents(): IManagedHoverContent | undefined {
-		const label = `${localize('chat.modelPicker.label', "Pick Model")}${super.getHoverContents()}`;
-		const { statusIcon, tooltip } = this.currentModel?.metadata || {};
+	override render(container: HTMLElement): void {
+		this._pickerWidget.render(container);
+		this.element = this._pickerWidget.domNode;
+		this._updateTooltip();
+		container.classList.add('chat-input-picker-item');
+	}
+
+	private _getAnchorElement(): HTMLElement {
+		if (this.element && getActiveWindow().document.contains(this.element)) {
+			return this.element;
+		}
+		return this.pickerOptions.getOverflowAnchor?.() ?? this.element!;
+	}
+
+	public openModelPicker(): void {
+		this._showPicker();
+	}
+
+	public show(): void {
+		this._showPicker();
+	}
+
+	public setEnabled(enabled: boolean): void {
+		this._pickerWidget.setEnabled(enabled);
+	}
+
+	private _showPicker(): void {
+		this._pickerWidget.show(this._getAnchorElement());
+	}
+
+	private _updateTooltip(): void {
+		if (!this.element) {
+			return;
+		}
+		const hoverContent = this._getHoverContents();
+		if (typeof hoverContent === 'string' && hoverContent) {
+			this._managedHover.value = getBaseLayerHoverDelegate().setupManagedHover(
+				getDefaultHoverDelegate('mouse'),
+				this.element,
+				hoverContent
+			);
+		} else {
+			this._managedHover.clear();
+		}
+	}
+
+	private _getHoverContents(): IManagedHoverContent | undefined {
+		let label = localize('chat.modelPicker.label', "Pick Model");
+		const keybindingLabel = this.keybindingService.lookupKeybinding(this._action.id, this._contextKeyService)?.getLabel();
+		if (keybindingLabel) {
+			label += ` (${keybindingLabel})`;
+		}
+		const { statusIcon, tooltip } = this._pickerWidget.selectedModel?.metadata || {};
 		return statusIcon && tooltip ? `${label} • ${tooltip}` : label;
 	}
-
-	protected override setAriaLabelAttributes(element: HTMLElement): void {
-		super.setAriaLabelAttributes(element);
-		const modelName = this.currentModel?.metadata.name ?? localize('chat.modelPicker.auto', "Auto");
-		element.ariaLabel = localize('chat.modelPicker.ariaLabel', "Pick Model, {0}", modelName);
-	}
-
-	protected override renderLabel(element: HTMLElement): IDisposable | null {
-		const { name, statusIcon } = this.currentModel?.metadata || {};
-		const domChildren = [];
-
-		if (statusIcon) {
-			const iconElement = renderIcon(statusIcon);
-			domChildren.push(iconElement);
-		}
-
-		domChildren.push(dom.$('span.chat-input-picker-label', undefined, name ?? localize('chat.modelPicker.auto', "Auto")));
-		domChildren.push(...renderLabelWithIcons(`$(chevron-down)`));
-
-		dom.reset(element, ...domChildren);
-		this.setAriaLabelAttributes(element);
-		return null;
-	}
-
 }
